requirements.txt
streamlit>=1.27.0

pdfplumber
python-docx
python-pptx
PyMuPDF

SpeechRecognition
pyaudio
pydub

transformers>=4.40.0
sentence-transformers
nltk
scikit-learn

optimum[openvino]>=1.16.1

torch>=2.0.0
torchvision
Pillow
opencv-python
tqdm

pytesseract

markdown
numpy
